Library Upgrades & Dependency Fixes (#846)
* Upgrading dependencies, some of which (Hazelcast and Coap) have small API changes.  Reducing CVE issue.
* Deprecating CoAP Connector

+++ kafka-connect-aws-s3/src/main/scala/io/lenses/streamreactor/connect/aws/s3/formats/ParquetFormatWriter.scala
@@ -61,7 +61,7 @@ class ParquetFormatWriter(outputStreamFn: () => S3OutputStream) extends S3Format
 
     AvroParquetWriter
       .builder[AnyRef](outputFile)
-      .withRowGroupSize(DEFAULT_BLOCK_SIZE)
+      .withRowGroupSize(DEFAULT_BLOCK_SIZE.toLong)
       .withPageSize(DEFAULT_PAGE_SIZE)
       .withSchema(schema)
       .build()
